Split up GitHub utility functions
