+++ src/utils/github/index.ts
@@ -1,0 +1,4 @@
+export * from './types';
+export { default as getRepoInfo } from './get-repo-info';
+export { default as getIssueInfo } from './get-issue-info';
+export { default as getPullInfo } from './get-pull-info';
